Doctor Appointment System
A Doctor Appointment system is a digital tool that allows patients to schedule appointments with healthcare providers online. The system typically integrates with the healthcare provider's existing schedules and provides patients with a user-friendly interface to choose an appointment date and time that works for them.
The system has 3 side's which are the administrator, doctor, and the patient.
By implementing these key features, a doctor appointment system can help healthcare providers streamline their appointment scheduling processes and provide patients with a more convenient and efficient way to schedule appointments